% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_partector.R
\name{plot_partector}
\alias{plot_partector}
\title{Plot Partector Data}
\usage{
plot_partector(
  d,
  yvar = "number",
  type = "line",
  crosshairs = TRUE,
  ymax = NULL,
  group_var = NULL,
  plot_alpha = 1
)
}
\arguments{
\item{d}{a dataframe created by \code{read_partector()}}

\item{yvar}{character; name of Partector measurement to plot. Default = 'number'.}

\item{type}{character; type of plot ('line', 'scatter', 'column'). Default = 'line'.}

\item{crosshairs}{logical; should a vertical crosshair be drawn when the cursor
is placed on the plot area?}

\item{ymax}{numeric; the maximum value of the y-axis. Default = NULL.}

\item{group_var}{character; name of grouping variable in the input data frame.
If specified, different series will be plotted for each group. Designating a
grouping variable can be useful if the input data frame includes data from multiple
units (i.e colocation test).}

\item{plot_alpha}{numeric; controls opacity of plot lines. Default = 1.}
}
\value{
a \emph{highchart}
}
\description{
\code{plot_partector} creates an interactive Highchart to visualize Partector data.
}
\examples{
\dontrun{

plot_partector <- function(d, yvar = 'number', type = 'line', crosshairs = TRUE,
                           ymax = NULL, group_var = NULL, plot_alpha = 1)
}


}
\seealso{
\url{https://jkunst.com/highcharter/}
}
